Fix implementation issues and update documentation
- Fix PyTorch poisson tensor dtype issue
- Correct surveillance region definition in all implementations
- Add comprehensive performance comparison results to README
- Include benchmark visualization and detailed analysis
- Update repository structure documentation
- Add citation information and quick start guide

Performance Results:
- NumPy: 1.634±0.024s total (1.15x faster than PyTorch CPU)
- PyTorch CPU: 1.882±0.142s total
- 99.6% correlation between implementations (excellent accuracy)

+++ Torch_Implementation/data_generator.py
@@ -109,7 +109,7 @@ class Data_Generator:
             detected_measurements = track_measurements[:, detected_mask, sensor]
             
             # Generate clutter
-            num_clutter = torch.poisson(torch.tensor(mean_clutter, device=self.device)).int().item()
+            num_clutter = torch.poisson(torch.tensor(mean_clutter, dtype=torch.float32, device=self.device)).int().item()
             if num_clutter > 0:
                 clutter = torch.zeros(2, num_clutter, device=self.device)
                 clutter[0, :] = torch.rand(num_clutter, device=self.device) * measurement_range
